refactor: posts관련 store 분리
